requirements.txt
streamlit==1.29.0
openai==1.3.5
pandas==2.1.3
fpdf==1.7.2
python-docx==1.0.0
pathlib==1.0.1
python-dotenv==1.0.0
